SK:
- BUG: statisticsonimage: floats were not properly converted to doubles by the ImageFileReader
- BUG: statisticsonimage: the itkStatisticsImageFilter returned a negative variance in case of homogeneous floating point image
- BUG: statisticsonimage: the screen output said that it was applying a mask when no mask was provided
- STY: some comments in the itkStatisticsImageFilterWithMask
- STY: sqrt->vcl_sqrt in itkStatisticsImageFilterWithMask (like in the cvs version of itk)

+++ src/statisticsonimage/statisticsonimage.cxx
@@ -89,17 +89,17 @@ int main( int argc, char ** argv )
   std::cout << "\tNumberOfComponents: " << NumberOfComponents << std::endl;
 
   /** force images to sneaky be converted to doubles */
-	ComponentType = "double";
+	ComponentType = "float";
 		
 	/** Run the program. */
 	try
 	{
-    run(StatisticsOnImage, double, 2, 1);
-    run(StatisticsOnImage, double, 2, 2);
-    run(StatisticsOnImage, double, 2, 3);
-    run(StatisticsOnImage, double, 3, 1);
-    run(StatisticsOnImage, double, 3, 2);
-    run(StatisticsOnImage, double, 3, 3);
+    run(StatisticsOnImage, float, 2, 1);
+    run(StatisticsOnImage, float, 2, 2);
+    run(StatisticsOnImage, float, 2, 3);
+    run(StatisticsOnImage, float, 3, 1);
+    run(StatisticsOnImage, float, 3, 2);
+    run(StatisticsOnImage, float, 3, 3);
   }
 	catch( itk::ExceptionObject &e )
 	{
